# swagger/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/healthcheck": {
      "get": {
        "operationId": "AppController_getHealthcheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/ingredient": {
      "post": {
        "operationId": "IngredientController_createIngredient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ingredient created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create ingredient",
        "tags": [
          "ingredients"
        ]
      },
      "get": {
        "operationId": "IngredientController_getIngredients",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all ingredients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get ingredients",
        "tags": [
          "ingredients"
        ]
      }
    },
    "/ingredient/{id}": {
      "put": {
        "operationId": "IngredientController_updateIngredient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIngredientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Ingredient updated."
          }
        },
        "summary": "Update ingredient",
        "tags": [
          "ingredients"
        ]
      },
      "delete": {
        "operationId": "IngredientController_removeIngredient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Ingredient removed."
          }
        },
        "summary": "Remove ingredient",
        "tags": [
          "ingredients"
        ]
      }
    },
    "/recipe": {
      "post": {
        "operationId": "RecipeController_createRecipe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create recipe",
        "tags": [
          "recipes"
        ]
      },
      "get": {
        "operationId": "getRecipes",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get recipes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get recipes",
        "tags": [
          "recipes"
        ]
      }
    },
    "/recipe/{id}": {
      "get": {
        "operationId": "RecipeController_getRecipeDetails",
        "parameters": [
          {
            "name": ":id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Get recipe by Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get recipe by Id",
        "tags": [
          "recipes"
        ]
      },
      "patch": {
        "operationId": "RecipeController_updateRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Recipe updated."
          }
        },
        "summary": "Update recipe",
        "tags": [
          "recipes"
        ]
      },
      "delete": {
        "operationId": "RecipeController_removeRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Recipe removed."
          }
        },
        "summary": "Remove recipe",
        "tags": [
          "recipes"
        ]
      }
    },
    "/meal": {
      "post": {
        "operationId": "MealController_createMeal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create meal",
        "tags": [
          "meals"
        ]
      },
      "get": {
        "operationId": "MealController_getMeals",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get meals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealListItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get meals",
        "tags": [
          "meals"
        ]
      }
    },
    "/meal/{id}": {
      "get": {
        "operationId": "MealController_getMealDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get meal details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get meal details",
        "tags": [
          "meals"
        ]
      },
      "patch": {
        "operationId": "MealController_updateMeal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Meal updated."
          }
        },
        "summary": "Update meal",
        "tags": [
          "meals"
        ]
      },
      "delete": {
        "operationId": "MealController_removeMeal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Meal removed."
          }
        },
        "summary": "Remove meal",
        "tags": [
          "meals"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "Auto-generated OpenAPI spec",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateIngredientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tomato"
          }
        },
        "required": [
          "name"
        ]
      },
      "IngredientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Tomato"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UpdateIngredientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tomato"
          }
        },
        "required": [
          "name"
        ]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "pageNum": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "total",
          "limit",
          "totalPages",
          "pageNum",
          "items"
        ]
      },
      "CreateRecipeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Egg soup"
          },
          "content": {
            "type": "string",
            "example": "Crack 2 eggs,..."
          },
          "img_url": {
            "type": "string",
            "nullable": true,
            "example": "imgSrc"
          },
          "ingredientIds": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "content",
          "img_url",
          "ingredientIds"
        ]
      },
      "RecipeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Egg soup"
          },
          "content": {
            "type": "string",
            "example": "- Dice tomato\n- Crack 2 eggs"
          },
          "img_url": {
            "type": "string",
            "example": "/imgSrc"
          },
          "ingredientIds": {
            "example": [
              "1",
              "2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "content",
          "img_url",
          "ingredientIds"
        ]
      },
      "RecipeListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Egg soup"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RecipeListResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "pageNum": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeListItemResponseDto"
            }
          }
        },
        "required": [
          "total",
          "limit",
          "totalPages",
          "pageNum",
          "items"
        ]
      },
      "UpdateRecipeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Egg soup"
          },
          "content": {
            "type": "string",
            "example": "Crack 2 eggs,..."
          },
          "img_url": {
            "type": "string",
            "nullable": true,
            "example": "imgSrc"
          },
          "ingredientIds": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "content",
          "img_url",
          "ingredientIds"
        ]
      },
      "CreateMealDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025/08/01"
          },
          "type": {
            "type": "string",
            "example": "lunch"
          },
          "recipe_ids": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "date",
          "type",
          "recipe_ids"
        ]
      },
      "MealResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "date": {
            "type": "string",
            "example": "2025-08-01"
          },
          "type": {
            "type": "string",
            "example": "lunch"
          },
          "recipe_ids": {
            "example": [
              "1",
              "2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "date",
          "type",
          "recipe_ids"
        ]
      },
      "MealItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "recipe_names": {
            "example": [
              "bun",
              "com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "recipe_names"
        ]
      },
      "MealListItemResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-08-01"
          },
          "lunch": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MealItemResponseDto"
              }
            ]
          },
          "dinner": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MealItemResponseDto"
              }
            ]
          }
        },
        "required": [
          "date",
          "lunch",
          "dinner"
        ]
      },
      "UpdateMealDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025/08/01"
          },
          "type": {
            "type": "string",
            "example": "lunch"
          },
          "recipe_ids": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "date",
          "type",
          "recipe_ids"
        ]
      }
    }
  }
}